Modify README example to not warn about duplicate names.
If the example usage of the cookbook is used, Chef will complain about
multiple resources with the same name. Since the name of the resource
is used as the chain name if not specified, just add an explicit chain
attribute for the resource and make the resource name more descriptive.

Fix issue #12.