Fix handling of `weight` attribute in simple_iptables_rule resource
Typically, the order of the rules in an iptables chain is determined by the
order of simple_iptables_rules resources in a recipe. However, if a non-standard
`weight` attribute value was used in *any* simple_iptables_rule, the relative
order of the rules that had the same weight became arbitrary. This is because
the code repeatedly calls `sort!` on the array of rules after adding every rule,
using the weights as the sort key, which potentially causes rules with the same
weight to be sorted in a different order with respect to each other on each
invocation of the `sort!` method, since they're viewed as equivalent.

Store rules in a hash where keys are weights and values are arrays of rules
having that weight. As new rules are added, they're appended to the relevant list.
When the final iptables rule file is generated, we output all the rules in
decreasing weight order.

Also take this opportunity to store rule in default-precedence attributes rather
than "normal" (persistent) attributes. As we reset all the attributes at the
beginning of the Chef run, there is no point to persisting the attributes.
This also resolves deprecation warnings for node.set in newer versions of Chef.

Resolves #81